Taking into account PR reviews
