--- iidm/iidm-mergingview/src/main/java/com/powsybl/iidm/mergingview/TerminalNodeBreakerViewAdapter.java
@@ -1,24 +1,0 @@
-/**
- * Copyright (c) 2019, RTE (http://www.rte-france.com)
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-package com.powsybl.iidm.mergingview;
-
-import com.powsybl.iidm.network.Terminal;
-
-public class TerminalNodeBreakerViewAdapter extends AbstractAdapter<Terminal.NodeBreakerView> implements Terminal.NodeBreakerView {
-
-    TerminalNodeBreakerViewAdapter(Terminal.NodeBreakerView delegate, MergingViewIndex index) {
-        super(delegate, index);
-    }
-
-    // -------------------------------
-    // Simple delegated methods ------
-    // -------------------------------
-    @Override
-    public int getNode() {
-        return getDelegate().getNode();
-    }
-}
